🦄 refactor(root)!: 将类型拆分出来。
